OHRM5X-823: Password changing feature related files moved to PIM plugin
