+++ symfony/plugins/orangehrmCorePlugin/config/routes.yaml
@@ -12,16 +12,3 @@ apiv2_core_about_organization:
   defaults:
     id: 0
     _api: OrangeHRM\Core\Api\Rest\AboutOrganizationAPI
-
-apiv2_core_update_password:
-  path: /api/v2/core/update-password
-  controller: OrangeHRM\Core\Controller\Rest\V2\GenericRestController::handle
-  methods: [ PUT ]
-  defaults:
-    id: 0
-    _api: OrangeHRM\Core\Api\Rest\UpdatePasswordAPI
-
-update_password:
-  path: /core/updatePassword
-  controller: OrangeHRM\Core\Controller\Common\UpdatePasswordController::handle
-  methods: [ GET ]
