+++ symfony/plugins/orangehrmPimPlugin/Controller/UpdatePasswordController.php
@@ -17,7 +17,7 @@
  * Boston, MA  02110-1301, USA
  */
 
-namespace OrangeHRM\Core\Controller\Common;
+namespace OrangeHRM\Pim\Controller;
 
 use OrangeHRM\Core\Controller\AbstractVueController;
 use OrangeHRM\Core\Traits\UserRoleManagerTrait;
